sent email to the producer, alter price_per_unit data type float4

+++ src/main/java/com/product/promotion/features/notice/Notice.java
@@ -50,6 +50,6 @@ public class Notice extends Audit implements Serializable {
     private Boolean alive;
 
     @Column(name="price_per_unit", nullable = false)
-    private Integer pricePerUnit;
+    private Float pricePerUnit;
 
 }
